+++ packages/renderer/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @vegastack/pages-renderer
 
+## 0.1.12
+
 ## 0.1.11
 
 ## 0.1.10